+++ packages/apps/components/CrossChain.tsx
@@ -17,18 +17,10 @@ import {
   CrossChainPayload,
   SubmitFn,
 } from 'shared/model';
-import {
-  isDarwinia2Ethereum,
-  isDVM2Substrate,
-  isEthereum2Darwinia,
-  isSubstrate2DVM,
-  isSubstrate2SubstrateDVM,
-  isSubstrateDVM2Substrate,
-} from 'shared/utils/bridge';
-import { emptyObsFactory, isKton, isRing } from 'shared/utils/helper';
-import { getDarwiniaBalance, getDVMBalance, getErc20Balance } from 'shared/utils/network/balance';
+import { emptyObsFactory, isRing } from 'shared/utils/helper';
 import { useAllowance } from '../hooks/allowance';
 import { useAccount, useApi } from '../providers';
+import { getBalance } from '../utils';
 import { BridgeSelector } from './form-control/BridgeSelector';
 import { Direction } from './form-control/Direction';
 import { FormItemButton } from './widget/FormItemButton';
@@ -39,47 +31,6 @@ const isDirectionChanged = (pre: CrossChainDirection, cur: CrossChainDirection) 
     { from: omit(cur.from, 'amount'), to: omit(cur.to, 'amount') }
   );
 };
-
-// eslint-disable-next-line complexity
-async function getBalance(direction: CrossChainDirection, account: string): Promise<BN[] | BN | null> {
-  const { from, to } = direction;
-
-  if (isEthereum2Darwinia(from.meta, to.meta)) {
-    const [ring, kton] = await Promise.all(
-      from.meta.tokens
-        .filter((item) => isRing(item.symbol) || isKton(item.symbol))
-        .sort((cur) => (isRing(cur.symbol) ? -1 : 1))
-        .map((item) => getErc20Balance(item.address, account, false))
-    );
-
-    return [ring, kton];
-  }
-
-  if (isDarwinia2Ethereum(from.meta, to.meta)) {
-    return getDarwiniaBalance(from.meta.provider, account);
-  }
-
-  if (isSubstrate2DVM(from.meta, to.meta)) {
-    return getDarwiniaBalance(from.meta.provider, account);
-  }
-
-  if (isDVM2Substrate(from.meta, to.meta)) {
-    const kton = from.meta.tokens.find((item) => item.type === 'native' && isKton(item.symbol))!;
-
-    return getDVMBalance(kton.address, account);
-  }
-
-  if (isSubstrate2SubstrateDVM(from.meta, to.meta)) {
-    return getDarwiniaBalance(from.meta.provider, account);
-  }
-
-  if (isSubstrateDVM2Substrate(from.meta, to.meta)) {
-    // from if xRing, xKton do not supported.
-    return getErc20Balance(from.address, account, false);
-  }
-
-  return null;
-}
 
 // eslint-disable-next-line complexity
 export function CrossChain({ dir }: { dir: CrossChainDirection }) {
@@ -110,14 +61,13 @@ export function CrossChain({ dir }: { dir: CrossChainDirection }) {
 
   const launch = useCallback(() => {
     form.validateFields().then((values) => {
-      console.log('💰 ~ cross chain args: ', values);
+      console.log('🚀 ~ cross chain args: ', values);
 
       if (!values.direction.from.amount) {
         message.error(t('Transfer amount is required'));
-        return;
+      } else {
+        submitFn(values);
       }
-
-      submitFn(values);
     });
   }, [form, submitFn, t]);
 
@@ -244,22 +194,7 @@ export function CrossChain({ dir }: { dir: CrossChainDirection }) {
         </Col>
 
         <Col xs={24} sm={{ span: 15, offset: 1 }} className="dark:bg-antDark">
-          <Form.Item
-            name={FORM_CONTROL.bridge}
-            rules={[
-              {
-                validator(_r, value: unknown) {
-                  if (!value) {
-                    message.error(t('Please select a bridge first!'));
-                    return Promise.reject();
-                  }
-
-                  return Promise.resolve();
-                },
-              },
-            ]}
-            className="mb-0"
-          >
+          <Form.Item name={FORM_CONTROL.bridge} rules={[{ required: true }]} className="mb-0">
             <BridgeSelector direction={direction} onChange={(value) => setBridge(value || null)} />
           </Form.Item>
         </Col>
