+++ packages/apps/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './balance';
 export * from './token';
